back/api/requirements.txt
uvicorn~=0.34.0
fastapi~=0.115.8
SQLAlchemy~=2.0.38
pydantic_settings
eva-decord
timm
einops
psycopg2-binary
faiss-cpu
sentence-transformers